simplified Voronoi constructor
moved creating the Rectangle to inside the constructor. no need to
create and pass one along.

had to make a second constructor - need to keep the same bounding
rectangle for each step of relaxation. otherwise, the border is not
drawn correctly

side note- it did look pretty cool though, kind of like a treasure
map...